+++ routes/api/contacts.js
@@ -1,91 +1,15 @@
 const express = require("express");
-const Joi = require("joi");
-
 const router = express.Router();
-const contacts = require("../../models/contacts");
-const { HttpError } = require("../../helpers");
+const controllers = require("../../controllers/contacts");
 
-const addContactSchema = Joi.object({
-  name: Joi.string().required().messages({
-    "string.base": `name should be a type of text`,
-    "any.required": `missing required name field`,
-  }),
-  email: Joi.string().email().required().messages({
-    "string.email": `email must be a valid email`,
-    "any.required": `missing required email field`,
-  }),
-  phone: Joi.string()
-    .pattern(/^[0-9()-]+$/)
-    .required()
-    .messages({
-      "string.pattern.base": `phone must contain only numbers, - and ( )`,
-      "any.required": `missing required phone field`,
-    }),
-});
+router.get("/", controllers.getAllContacts);
 
-router.get("/", async (req, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/:contactId", controllers.getContactById);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.getContactById(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", controllers.addContact);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = addContactSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", controllers.removeContact);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const { contactId } = req.params;
-    const result = await contacts.removeContact(contactId);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json({ message: "Contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const { error } = addContactSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
-    const { contactId } = req.params;
-    const result = await contacts.updateContact(contactId, req.body);
-    if (!result) {
-      throw HttpError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", controllers.updateContact);
 
 module.exports = router;
